feat: 多平台去水印解析 + 视频代理 + 基础设施修复
- 新增 media-parser 集成，支持 24 平台去水印解析
- 新增视频/图片代理端点，按域名动态设 Referer 绕过防盗链
- DYVideoServer: URL 自动提取 + mediaType 判定(video/image/mixed)
- ApiResponseFilter 修复泛型判断，防止 ApiResponse 双重包装
- ApiResponse.SuccessResponse 合并重载 + 默认 Code=200
- Program.cs 注册 IMediaParserService HttpClient
- .gitignore 清理: 排除 bin/obj/logs/publish/.vs/node_modules
- 新增 README.md 项目说明
- 代码小修: UserService 异常类型, AuthController 参数类型, PigeonController 新增, appsettings HTTPS 配置

+++ Controllers/AuthController.cs
@@ -95,7 +95,7 @@ namespace day1.Controllers
 
         [Authorize(Roles = "Admin")]
         [HttpDelete("DeleteByUserId")]
-        public IActionResult DeleteByUserId(int userid)
+        public IActionResult DeleteByUserId(long userid)
         {
             _userService.DeleteByUserId(userid);
             return Ok("删除成功!");
